+++ src/ESFA.DC.ILR.FundingService.FM35.ExternalData/Interface/IReferenceDataCachePopulationService.cs
@@ -4,6 +4,6 @@ namespace ESFA.DC.ILR.FundingService.FM35.ExternalData.Interface
 {
     public interface IReferenceDataCachePopulationService
     {
-        void Populate(IList<string> learnAimRefs, IList<string> postcodes, IList<int> orgUkprns, IList<int> lEmpIDs);
+        void Populate(IList<string> learnAimRefs, IList<string> postcodes, IList<long> orgUkprns, IList<int> lEmpIDs);
     }
 }
